+++ test/Laplacian/HodgeLaplacian_vector.jl
@@ -92,10 +92,7 @@ function hodge_laplacian_vector(
   T = TestFESpace(Ω_atlas, ReferenceFE(lagrangian,Float64,p_fe+1); conformity=:H1)
   S = TrialFESpace(T)
 
-  # For Gridap 0.19, we can use the standard nedelec constructor. That is, we would use:
-  # R = TestFESpace(Ω_atlas, ReferenceFE(nedelec,Float64,p_fe);conformity=:Hcurl)
-  # For Gridap 0.20, we need to set change_dof=false due to the new definition of dofs
-  R = TestFESpace(Ω_atlas, ReferenceFE(nedelec,Float64,p_fe; change_dof=false);conformity=:Hcurl)
+  R = TestFESpace(Ω_atlas, ReferenceFE(nedelec,Float64,p_fe); conformity=:Hcurl)
   H = TrialFESpace(R)
 
   sigma_int = interpolate(sigma_cf,S)